fix: Rate Payloads成獨立物件 && 重構共同邏輯
